pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ontograph"
version = "0.1.0"
description = "Aerospace document ingestion, ontology mapping, synthesis, and impact evaluation pipeline"
requires-python = ">=3.11"
dependencies = [
    # document conversion
    "pymupdf>=1.24",
    "markdown-it-py>=3.0",
    # ontology
    "rdflib>=7.0",
    # LLM providers
    "anthropic>=0.30",
    "openai>=1.40",
    "google-generativeai>=0.7",
    # data validation
    "pydantic>=2.7",
    # embeddings / vector search
    "numpy>=1.26",
    # CLI
    "rich>=13.7",
    "typer>=0.12",
    # env loading
    "python-dotenv>=1.0",
    # YAML (org-knowledge authoring)
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
ontograph-ingest    = "cli.ingest:app"
ontograph-synthesize = "cli.synthesize:app"
ontograph-evaluate  = "cli.evaluate:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -q"
